pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "bijoux-router"
version = "0.1.0"
description = "Bijoux Quota-Aware LLM Gateway — multi-provider LLM router with quota tracking, failover, and persistence"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.11"
authors = [
    {name = "Bijoux Team"},
]
keywords = ["llm", "router", "gateway", "quota", "openrouter", "gemini", "openai"]

dependencies = [
    "pydantic>=2.5,<3",
    "httpx>=0.25,<1",
    "pyyaml>=6.0,<7",
    "click>=8.1,<9",
    "tiktoken>=0.5,<1",
    "tenacity>=8.2,<10",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4",
    "pytest-asyncio>=0.23",
    "pytest-cov>=4.1",
    "respx>=0.21",
    "ruff>=0.1",
    "mypy>=1.7",
]

[project.scripts]
bijoux = "bijoux_router.cli.main:cli"

[tool.setuptools.packages.find]
include = ["bijoux_router*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
markers = [
    "integration: marks tests as integration tests",
    "slow: marks tests as slow",
]

[tool.ruff]
target-version = "py311"
line-length = 120

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_configs = true
